backend/requirements.txt
flask
flask-cors
torch
transformers
Pillow
requests
numpy
gunicorn
